requirements.txt
streamlit>=1.28.0
yt-dlp>=2023.10.13
python-dotenv>=1.0.0
requests>=2.31.0
imageio-ffmpeg>=0.4.9
